updating the signature_v to accomate the 64 bit chainId as well

+++ python/src/trezorlib/messages.py
@@ -3975,7 +3975,7 @@ class EthereumTxRequest(protobuf.MessageType):
     MESSAGE_WIRE_TYPE = 59
     FIELDS = {
         1: protobuf.Field("data_length", "uint32", repeated=False, required=False),
-        2: protobuf.Field("signature_v", "uint32", repeated=False, required=False),
+        2: protobuf.Field("signature_v", "uint64", repeated=False, required=False),
         3: protobuf.Field("signature_r", "bytes", repeated=False, required=False),
         4: protobuf.Field("signature_s", "bytes", repeated=False, required=False),
     }
